feat(client): adds basic chat component
