+++ packages/client/src/components/Group/components/Chat/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Chat';
