+++ packages/client/src/components/Group/components/VideoView/VideoView.tsx
@@ -1,10 +1,11 @@
 import { Box } from '@material-ui/core';
-import React, { useState } from 'react';
+import React from 'react';
 
 import useIdleTimer from '../../../../lib/hooks/useIdleTImer';
 import { User } from '../../lib/useSocketHandler';
 import { Message } from '../../lib/useSocketHandler/lib/useChatMessages';
 import Activities from '../Activities';
+import Chat from '../Chat';
 import VideoControls from '../VideoControls';
 import * as S from './VideoView.styles';
 
@@ -31,18 +32,6 @@ interface Props {
 
 export default function VideoView(props: Props) {
   const { isIdle } = useIdleTimer({ wait: 3500 });
-  const [newChatMessage, setNewChatMessage] = useState('');
-
-  const submitChatMessage = (e: React.SyntheticEvent) => {
-    e.preventDefault();
-    console.log(e);
-    props.sendMessage(newChatMessage);
-    setNewChatMessage('');
-  };
-
-  const onChatMessageChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setNewChatMessage(e.target.value);
-  };
 
   return (
     <S.VideoView
@@ -65,18 +54,8 @@ export default function VideoView(props: Props) {
           />
         </Box>
       )}
+      <Chat messages={props.messages} sendMessage={props.sendMessage} />
       <Box height="100%">{props.children({ streams: props.streams })}</Box>
-      <div>
-        {props.messages.map(({ message }) => {
-          return <div>{message}</div>;
-        })}
-
-        <div>
-          <form onSubmit={submitChatMessage}>
-            <input value={newChatMessage} onChange={onChatMessageChange} />
-          </form>
-        </div>
-      </div>
 
       <S.Controls isIdle={isIdle}>
         <VideoControls
